requirements.txt
langchain>=0.2.0
langgraph>=0.0.40
langchain-openai>=0.1.0
python-dotenv>=1.0.0
requests>=2.0.0
beautifulsoup4>=4.0.0
fastapi>=0.100.0
uvicorn>=0.20.0